+++ lib/components/collection_views/ad_list_view/widgets/dismissible_ad.dart
@@ -17,7 +17,7 @@
 
 import 'package:flutter/material.dart';
 
-import '/features/my_account/my_ads/model/my_ads_dismissible.dart';
+import '../../../../features/account/my_ads/model/my_ads_dismissible.dart';
 import '/core/models/ad.dart';
 import '../../../widgets/base_dismissible_container.dart';
 import 'ad_card_view.dart';
