+++ lib/features/filters/filters_screen.dart
@@ -20,7 +20,7 @@ import 'package:flutter/material.dart';
 import '../../core/models/ad.dart';
 import '../../core/models/filter.dart';
 import '../../components/buttons/big_button.dart';
-import '../my_account/mechanics/mechanics_screen.dart';
+import '../account/mechanics/mechanics_screen.dart';
 import 'filters_controller.dart';
 import 'filters_states.dart';
 import 'widgets/text_form_dropdown.dart';
